+++ app/src/main/java/com/example/playlistmaker/Constants.kt
@@ -1,7 +1,5 @@
 package com.example.playlistmaker
 
-import android.os.Parcelable
-
 object Constants {
     const val THEME_PREF_KEY = "theme_sharedprefs_key"
     const val RECENT_TRACKS_KEY = "recent_list_sharedprefs_key"
