+++ app/src/main/java/com/example/playlistmaker/SearchAdapter.kt
@@ -2,24 +2,18 @@ package com.example.playlistmaker
 
 
 import android.content.Intent
-import android.util.TypedValue
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
-import android.widget.ImageView
-import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
-import com.bumptech.glide.Glide
-import com.bumptech.glide.load.resource.bitmap.CenterCrop
-import com.bumptech.glide.load.resource.bitmap.RoundedCorners
-import com.example.playlistmaker.domain.datamodels.Track
-import com.example.playlistmaker.presentation.audioplayer.PlayerActivity
+
+import com.example.playlistmaker.audioplayer.domain.datamodels.Track
+import com.example.playlistmaker.audioplayer.ui.audioplayer.PlayerActivity
 import java.util.ArrayList
 
 class SearchAdapter(diplayedList: ArrayList<Track>, activity: SearchActivity) :
 
 
-    RecyclerView.Adapter<SearchAdapter.TrackViewHolder>() {
+    RecyclerView.Adapter<TrackViewHolder>() {
     val list = diplayedList
     val activityInstance = activity
     val debounce = Debounce()
@@ -49,32 +43,6 @@ class SearchAdapter(diplayedList: ArrayList<Track>, activity: SearchActivity) :
     }
 
     override fun getItemCount(): Int = list.size
-
-    class TrackViewHolder(itemView: View) :
-        RecyclerView.ViewHolder(itemView) {
-        private val trackName: TextView = itemView.findViewById(R.id.track_name)
-        private val artistName: TextView = itemView.findViewById(R.id.artist_name)
-        private val trackTime: TextView = itemView.findViewById(R.id.track_time)
-        private val artworkImageView: ImageView = itemView.findViewById(R.id.track_artwork)
-
-        fun bind(track: Track) {
-            trackName.text = track.trackName
-            artistName.text = track.artistName
-            trackTime.text = track.trackTime
-            val roundingRadius = itemView.resources.getDimension(R.dimen.Track_icon_rounding)
-            val pixelsForRoundedCorners = TypedValue.applyDimension(
-                TypedValue.COMPLEX_UNIT_PX,
-                roundingRadius,
-                itemView.resources.displayMetrics
-            )
-            Glide.with(artworkImageView)
-                .load(track.artworkUrl100)
-                .placeholder(R.drawable.placeholder)
-                .transform(CenterCrop(), RoundedCorners(pixelsForRoundedCorners.toInt()))
-                .into(artworkImageView)
-        }
-
-    }
 
 }
