+++ app/src/main/java/com/example/playlistmaker/TrackViewHolder.kt
@@ -1,0 +1,37 @@
+package com.example.playlistmaker
+
+import android.util.TypedValue
+import android.view.View
+import android.widget.ImageView
+import android.widget.TextView
+import androidx.recyclerview.widget.RecyclerView
+import com.bumptech.glide.Glide
+import com.bumptech.glide.load.resource.bitmap.CenterCrop
+import com.bumptech.glide.load.resource.bitmap.RoundedCorners
+import com.example.playlistmaker.audioplayer.domain.datamodels.Track
+
+class TrackViewHolder(itemView: View) :
+    RecyclerView.ViewHolder(itemView) {
+    private val trackName: TextView = itemView.findViewById(R.id.track_name)
+    private val artistName: TextView = itemView.findViewById(R.id.artist_name)
+    private val trackTime: TextView = itemView.findViewById(R.id.track_time)
+    private val artworkImageView: ImageView = itemView.findViewById(R.id.track_artwork)
+
+    fun bind(track: Track) {
+        trackName.text = track.trackName
+        artistName.text = track.artistName
+        trackTime.text = track.trackTime
+        val roundingRadius = itemView.resources.getDimension(R.dimen.Track_icon_rounding)
+        val pixelsForRoundedCorners = TypedValue.applyDimension(
+            TypedValue.COMPLEX_UNIT_PX,
+            roundingRadius,
+            itemView.resources.displayMetrics
+        )
+        Glide.with(artworkImageView)
+            .load(track.artworkUrl100)
+            .placeholder(R.drawable.placeholder)
+            .transform(CenterCrop(), RoundedCorners(pixelsForRoundedCorners.toInt()))
+            .into(artworkImageView)
+    }
+
+}
