+++ app/src/main/java/com/example/playlistmaker/audioplayer/data/MediaPlayerState.kt
@@ -1,0 +1,12 @@
+package com.example.playlistmaker.audioplayer.data
+
+enum class MediaPlayerState {
+    IDLE,
+    INITIALIZED,
+    PREPARED,
+    STARTED,
+    PAUSED,
+    STOPPED,
+    PLAYBACK_COMPLETED,
+    ERROR
+}
