+++ app/src/main/java/com/example/playlistmaker/audioplayer/data/network/RetrofitApiClient.kt
@@ -1,9 +1,9 @@
-package com.example.playlistmaker
+package com.example.playlistmaker.audioplayer.data.network
 
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 
-class ApiClient {
+class RetrofitApiClient {
     val retrofit = Retrofit.Builder()
         .baseUrl("https://itunes.apple.com")
         .addConverterFactory(GsonConverterFactory.create())
