+++ app/src/main/java/com/example/playlistmaker/audioplayer/domain/datamodels/Track.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker.domain.datamodels
+package com.example.playlistmaker.audioplayer.domain.datamodels
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
