--- app/src/main/java/com/example/playlistmaker/domain/searchServerResponse.kt
@@ -1,8 +1,0 @@
-package com.example.playlistmaker.domain
-
-import com.example.playlistmaker.domain.datamodels.unparsedTrack
-
-data class searchServerResonse(
-    val resultCount:Int,
-    val results: List<unparsedTrack>
-)
