+++ app/src/main/java/com/example/playlistmaker/searchServerResponse.kt
@@ -1,0 +1,8 @@
+package com.example.playlistmaker
+
+import com.example.playlistmaker.audioplayer.data.dto.unparsedTrack
+
+data class searchServerResonse(
+    val resultCount:Int,
+    val results: List<unparsedTrack>
+)
